requirements.txt
Flask==2.2.2
flask-restx==1.0.3
PyJWT==2.6.0